Reintegrate schema_exposure into core
