+++ codefeedr-core/src/main/scala/org/codefeedr/buffer/serialization/schema_exposure/RedisSchemaExposer.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  *
  */
-package org.codefeedr.util.schema_exposure
+package org.codefeedr.buffer.serialization.schema_exposure
 
 import java.net.URI
