+++ codefeedr-core/src/test/scala/org/codefeedr/buffer/serialization/schema_exposure/SchemaExposerTest.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  *
  */
-package org.codefeedr.util.schema_exposure
+package org.codefeedr.buffer.serialization.schema_exposure
 
 import org.apache.avro.Schema
 import org.scalatest.{BeforeAndAfter, FunSuite}
