+++ codefeedr-core/src/test/scala/org/codefeedr/buffer/serialization/schema_exposure/ZookeeperSchemaExposerTest.scala
@@ -16,30 +16,25 @@
  * limitations under the License.
  *
  */
-package org.codefeedr.util.schema_exposure
+package org.codefeedr.buffer.serialization.schema_exposure
 
-import com.github.sebruck.EmbeddedRedis
+import net.manub.embeddedkafka.{EmbeddedKafka, EmbeddedKafkaConfig}
 import org.scalatest.BeforeAndAfterAll
-import redis.embedded.RedisServer
 
-class RedisSchemaExposerTest
+class ZookeeperSchemaExposerTest
   extends SchemaExposerTest
-    with BeforeAndAfterAll
-    with EmbeddedRedis {
-  var redis: RedisServer = _
-  var redisPort: Int = 0
+    with EmbeddedKafka
+    with BeforeAndAfterAll {
 
-  // Before all tests, setup an embedded redis
   override def beforeAll(): Unit = {
-    redis = startRedis()
-    redisPort = redis.ports().get(0)
+    implicit val config: EmbeddedKafkaConfig = EmbeddedKafkaConfig(zooKeeperPort = 2181)
+    EmbeddedKafka.start()
   }
 
-  // After all tests, stop embedded redis
   override def afterAll(): Unit = {
-    stopRedis(redis)
+    EmbeddedKafka.stop()
   }
 
   override def getSchemaExposer: SchemaExposer =
-    new RedisSchemaExposer(s"redis://localhost:$redisPort")
+    new ZookeeperSchemaExposer("localhost:2181")
 }
